[ADD] l10n_in_edi_irn_ewaybill: separate irn from ewaybill module
Before this commit
==================
The functionality of the E-waybill based on IRN (Invoice Reference Number) was
integrated within the E-waybill module `l10n_in_edi_ewaybill`.

After this commit
=================
This commit separates the functionality of the E-waybill based on IRN from the
E-waybill module in `l10n_in_edi_irn_ewaybill`.
