+++ addons/l10n_in_edi_irn_ewaybill/__manifest__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+{
+    "name": "Indian - IRN E-waybill",
+    "countries": ["in"],
+    "version": "1.0",
+    "category": "Accounting/Localizations/EDI",
+    "depends": [
+        "l10n_in_edi_ewaybill",
+    ],
+    "installable": True,
+    "auto_install": True,
+    "license": "LGPL-3",
+}
